+++ apps/web/src/app/customer/proposals/[reportId]/_components/consult-action-style.ts
@@ -1,0 +1,2 @@
+export const ACTION_BUTTON_CLASS =
+  "h-[2.125rem] px-[0.9375rem] py-[0.5625rem] text-[0.8125rem]";
